% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{forward}
\alias{forward}
\title{function for computing forward probabilities}
\usage{
forward(initial_probs, transition_matrix, emission_matrix, observations)
}
\arguments{
\item{initial_probs}{1xN vector of initial probabilities}

\item{transition_matrix}{NxN matrix of hidden state transition probabilities}

\item{emission_matrix}{NxK matrix of emission probabilities}

\item{observations}{vector of observed variable observations}
}
\value{
A NxT matrix of forward probabilities
}
\description{
function for computing forward probabilities
}
