% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{forward_algorithm}
\alias{forward_algorithm}
\title{Forward algorithm for likelihood computation}
\usage{
forward_algorithm(
  initial_probs,
  transition_matrix,
  emission_matrix,
  observations
)
}
\arguments{
\item{initial_probs}{1xN vector of initial probabilities}

\item{transition_matrix}{NxN matrix of hidden state transition probabilities}

\item{emission_matrix}{NxK matrix of emission probabilities}

\item{observations}{vector of observed variable observations}
}
\value{
Model likelihood for given data
}
\description{
Forward algorithm for likelihood computation
}
