% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{forward_algorithm_tensor}
\alias{forward_algorithm_tensor}
\title{Forward algorithm for likelihood computation with multiple emission and transition matrices}
\usage{
forward_algorithm_tensor(
  initial_probs,
  transition_matrix,
  emission_matrix,
  observations,
  bin
)
}
\arguments{
\item{initial_probs}{1xN vector of initial probabilities}

\item{transition_matrix}{NxNxB cube of hidden state transition probabilities}

\item{emission_matrix}{NxKxB cube of emission probabilities}

\item{observations}{vector of observed variable observations}

\item{bin}{which bin generated each observation (pair of emission and transition matrices)}
}
\value{
Model likelihood for given data
}
\description{
Forward algorithm for likelihood computation with multiple emission and transition matrices
}
