% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_hmm.R
\name{simulate_hmm}
\alias{simulate_hmm}
\title{simulate_hmm}
\usage{
simulate_hmm(pi, A, B, n_obs)
}
\arguments{
\item{pi}{initial state probabilities}

\item{A}{matrix of transition probabilities}

\item{B}{matrix of emission probabilities}

\item{n_obs}{Number of observations to simulate}
}
\value{
List with the sequence of observed variables (x) and underlying hidden states (z)
}
\description{
Simulate a hidden Markov Model
}
