% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_parameters}
\alias{update_parameters}
\title{function for computing updated HMM parameters}
\usage{
update_parameters(
  initial_probs,
  transition_matrix,
  emission_matrix,
  observations
)
}
\arguments{
\item{initial_probs}{1xN vector of initial probabilities}

\item{transition_matrix}{NxN matrix of hidden state transition probabilities}

\item{emission_matrix}{NxK matrix of emission probabilities}

\item{observations}{vector of observed variable observations}
}
\value{
list with updated parameters
}
\description{
function for computing updated HMM parameters
}
