% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{viterbi_}
\alias{viterbi_}
\alias{viterbi}
\title{Viterbi algorithm}
\usage{
viterbi(initial_probs, transition_matrix, emission_probabilities, observations)
}
\arguments{
\item{initial_probs}{1xN vector of initial probabilities}

\item{transition_matrix}{NxN matrix of hidden state transition probabilities}

\item{emission_probabilities}{NxK matrix of emission probabilities}

\item{observations}{vector of observed variable observations}
}
\value{
most likely sequence of hidden states
}
\description{
Viterbi algorithm
}
